pyproject.toml
[project]
name = "your-project-name"
version = "1.0.0"
description = "Your project description"
requires-python = ">=3.11,<3.14"

dependencies = [
    "accelerate",
    "aiofiles",
    "aiohttp",
    "anndata",
    "av",
    "beautifulsoup4",
    "bitsandbytes>=0.49",
    "click",
    "colorama",
    "coloredlogs",
    "datasets",
    "diffusers",
    "docx2txt",
    "einops",
    "einx",
    "emoji",
    "encodec",
    "gTTS",
    "h5py",
    "httpx",
    "httpx-sse",
    "huggingface-hub",
    "HyperPyYAML",
    "img2pdf",
    "langchain",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-huggingface>=0.3.1",
    "langchain-text-splitters>=0.3.11",
    "langdetect>=1.0",
    "llvmlite",
    "lxml",
    "Markdown",
    "msg-parser",
    "natsort",
    "networkx",
    "nltk",
    "numba",
    "numpy",
    "olefile",
    "openai",
    "openpyxl",
    "optimum",
    "pandas",
    "pillow",
    "pipdeptree",
    "protobuf",
    "psutil",
    "pyarrow",
    "pydantic",
    "pydantic-settings",
    "pymupdf",
    "pypandoc",
    "pypdf",
    "pyreadline3",
    "pyside6",
    "python-docx",
    "python-dotenv",
    "python-magic",
    "PyYAML",
    "rapidfuzz",
    "regex",
    "requests",
    "rich",
    "ruamel.yaml",
    "safetensors",
    "scikit-learn",
    "scipy",
    "sentence-transformers",
    "sentencepiece",
    "sounddevice",
    "soundfile",
    "speechbrain",
    "SQLAlchemy",
    "tabulate2",
    "termcolor",
    "tiktoken",
    "tiledb",
    "tiledb-cloud",
    "tiledb-vector-search",
    "timm",
    "tokenizers",
    "tqdm",
    "transformers",
    "unstructured",
    "unstructured-client",
    "vector-quantize-pytorch",
    "vocos",
    "watchdog",
    "xlrd",
]

[project.optional-dependencies]
gpu = [
    # PyTorch - Python 3.11
    "torch @ https://download.pytorch.org/whl/cu128/torch-2.9.0%2Bcu128-cp311-cp311-win_amd64.whl ; python_version == '3.11'",
    "torchvision @ https://download.pytorch.org/whl/cu128/torchvision-0.24.0%2Bcu128-cp311-cp311-win_amd64.whl ; python_version == '3.11'",
    "torchaudio @ https://download.pytorch.org/whl/cu128/torchaudio-2.9.0%2Bcu128-cp311-cp311-win_amd64.whl ; python_version == '3.11'",
    "flash-attn @ https://github.com/kingbri1/flash-attention/releases/download/v2.8.3/flash_attn-2.8.3+cu128torch2.9.0cxx11abiFALSE-cp311-cp311-win_amd64.whl ; python_version == '3.11'",
    
    # PyTorch - Python 3.12
    "torch @ https://download.pytorch.org/whl/cu128/torch-2.9.0%2Bcu128-cp312-cp312-win_amd64.whl ; python_version == '3.12'",
    "torchvision @ https://download.pytorch.org/whl/cu128/torchvision-0.24.0%2Bcu128-cp312-cp312-win_amd64.whl ; python_version == '3.12'",
    "torchaudio @ https://download.pytorch.org/whl/cu128/torchaudio-2.9.0%2Bcu128-cp312-cp312-win_amd64.whl ; python_version == '3.12'",
    "flash-attn @ https://github.com/kingbri1/flash-attention/releases/download/v2.8.3/flash_attn-2.8.3+cu128torch2.9.0cxx11abiFALSE-cp312-cp312-win_amd64.whl ; python_version == '3.12'",
    
    # PyTorch - Python 3.13
    "torch @ https://download.pytorch.org/whl/cu128/torch-2.9.0%2Bcu128-cp313-cp313-win_amd64.whl ; python_version == '3.13'",
    "torchvision @ https://download.pytorch.org/whl/cu128/torchvision-0.24.0%2Bcu128-cp313-cp313t-win_amd64.whl ; python_version == '3.13'",
    "torchaudio @ https://download.pytorch.org/whl/cu128/torchaudio-2.9.0%2Bcu128-cp313-cp313-win_amd64.whl ; python_version == '3.13'",
    "flash-attn @ https://github.com/kingbri1/flash-attention/releases/download/v2.8.3/flash_attn-2.8.3+cu128torch2.9.0cxx11abiFALSE-cp313-cp313-win_amd64.whl ; python_version == '3.13'",
    
    # NVIDIA packages
    "nvidia-cuda-runtime-cu12==12.8.90",
    "nvidia-cublas-cu12==12.8.4.1",
    "nvidia-cuda-nvrtc-cu12==12.8.93",
    "nvidia-cuda-nvcc-cu12==12.8.93",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-ml-py==13.580.82",
    
    # Other GPU packages
    "triton-windows==3.5.1.post24",
    "xformers==0.0.33.post1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []